pyproject.toml
[project]
name = "fast-porject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "deptry>=0.23.0",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "mypy>=1.15.0",
    "passlib[bcrypt]>=1.7.4",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.3",
    "pyjwt>=2.10.1",
    "ruff>=0.11.8",
    "semgrep>=1.121.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]
